third_party/blink/web_tests/platform/generic/fast/forms/image/input-width-height-attributes-without-renderer-loaded-image-expected.txt

FAIL e.width should be 16. Was 0.
PASS e.height is 16
PASS successfullyParsed is true

TEST COMPLETE
